changed type because we don't need that much precision, and float is json serializable.

+++ backend/app/schemas/observation_schema.py
@@ -1,24 +1,22 @@
 from enum import Enum
 
-from decimal import Decimal
-
 from pydantic import BaseModel, Field
 
 
 class ObservationModel(BaseModel):
     """Info required to query the ml model."""
-    sepal_length: Decimal = Field(
+    sepal_length: float = Field(
         description= "Sepal Length (cm)",
-        gt=0, le=11, decimal_places=2)
-    sepal_width: Decimal = Field(
+        gt=0, le=11)
+    sepal_width: float = Field(
         description="Sepal Width (cm)",
-        gt=0, le=6, decimal_places=2)
-    petal_length: Decimal = Field(
+        gt=0, le=6)
+    petal_length: float = Field(
         description="Petal Length (cm)",
-        gt=0, le=10, decimal_places=2)
-    petal_width: Decimal = Field(
+        gt=0, le=10)
+    petal_width: float = Field(
         description="Petal Width (cm)",
-        gt=0, le=4, decimal_places=2)
+        gt=0, le=4)
 
 
 class SpeciesEnum(str, Enum):
